+++ src/main/java/com/ssafyss/board_practice/auth/constants/SuccessMessages.java
@@ -1,0 +1,7 @@
+package com.ssafyss.board_practice.auth.constants;
+
+public class SuccessMessages {
+    public static final String EMAIL_AVAILABLE = "사용할 수 있는 이메일입니다";
+    public static final String SIGN_UP_SUCCESS = "회원가입 성공";
+    public static final String SIGN_IN_SUCCESS = "로그인 성공";
+}
