+++ src/main/java/com/ssafyss/board_practice/auth/dto/CheckEmailRequest.java
@@ -1,0 +1,13 @@
+package com.ssafyss.board_practice.auth.dto;
+
+public class CheckEmailRequest {
+    private String email;
+
+    public CheckEmailRequest() {
+
+    }
+
+    public String getEmail() {
+        return email;
+    }
+}
