+++ src/main/java/com/ssafyss/board_practice/auth/dto/CreateUserRequest.java
@@ -1,0 +1,20 @@
+package com.ssafyss.board_practice.auth.dto;
+
+public class CreateUserRequest {
+    private String email;
+
+    private String password;
+
+    public CreateUserRequest() {
+
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+}
