+++ src/main/java/com/ssafyss/board_practice/auth/exception/DuplicatedEmailException.java
@@ -1,0 +1,7 @@
+package com.ssafyss.board_practice.auth.exception;
+
+public class DuplicatedEmailException extends RuntimeException {
+    public DuplicatedEmailException(String message) {
+        super(message);
+    }
+}
