+++ src/main/java/com/ssafyss/board_practice/auth/exception/SignInFailedException.java
@@ -1,0 +1,7 @@
+package com.ssafyss.board_practice.auth.exception;
+
+public class SignInFailedException extends RuntimeException {
+    public SignInFailedException(String message) {
+        super(message);
+    }
+}
