--- src/main/java/com/ssafyss/board_practice/domain/auth/constants/ErrorMessages.java
@@ -1,6 +1,0 @@
-package com.ssafyss.board_practice.domain.auth.constants;
-
-public class ErrorMessages {
-    public static final String EMAIL_DUPLICATED = "중복된 이메일입니다";
-    public static final String SIGN_IN_FAILED = "이메일 혹은 비밀번호가 틀립니다";
-}
