--- src/main/java/com/ssafyss/board_practice/domain/auth/dto/CheckEmailRequest.java
@@ -1,13 +1,0 @@
-package com.ssafyss.board_practice.domain.auth.dto;
-
-public class CheckEmailRequest {
-    private String email;
-
-    public CheckEmailRequest() {
-
-    }
-
-    public String getEmail() {
-        return email;
-    }
-}
