--- src/main/java/com/ssafyss/board_practice/domain/auth/exception/DuplicatedEmailException.java
@@ -1,7 +1,0 @@
-package com.ssafyss.board_practice.domain.auth.exception;
-
-public class DuplicatedEmailException extends RuntimeException {
-    public DuplicatedEmailException(String message) {
-        super(message);
-    }
-}
