--- src/main/java/com/ssafyss/board_practice/domain/auth/exception/SignInFailedException.java
@@ -1,7 +1,0 @@
-package com.ssafyss.board_practice.domain.auth.exception;
-
-public class SignInFailedException extends RuntimeException {
-    public SignInFailedException(String message) {
-        super(message);
-    }
-}
