--- src/main/java/com/ssafyss/board_practice/domain/todo/dao/TodoDao.java
@@ -1,18 +1,0 @@
-package com.ssafyss.board_practice.domain.todo.dao;
-
-import com.ssafyss.board_practice.domain.todo.dto.ReadTodoResponse;
-import com.ssafyss.board_practice.domain.todo.dto.UpdateTodoRequest;
-import com.ssafyss.board_practice.domain.todo.entity.Todo;
-import java.util.List;
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface TodoDao {
-    void insert(Todo todo);
-
-    List<ReadTodoResponse> findByUserId(Long userId);
-
-    int deleteById(Long id);
-
-    void update(UpdateTodoRequest request);
-}
