--- src/main/java/com/ssafyss/board_practice/domain/todo/dto/CreateTodoRequest.java
@@ -1,9 +1,0 @@
-package com.ssafyss.board_practice.domain.todo.dto;
-
-import lombok.Getter;
-
-@Getter
-public class CreateTodoRequest {
-    private Long userId;
-    private String content;
-}
