--- src/main/java/com/ssafyss/board_practice/domain/todo/dto/DeleteTodoRequest.java
@@ -1,8 +1,0 @@
-package com.ssafyss.board_practice.domain.todo.dto;
-
-import lombok.Getter;
-
-@Getter
-public class DeleteTodoRequest {
-    private Long id;
-}
