--- src/main/java/com/ssafyss/board_practice/domain/todo/dto/ReadTodoRequest.java
@@ -1,8 +1,0 @@
-package com.ssafyss.board_practice.domain.todo.dto;
-
-import lombok.Getter;
-
-@Getter
-public class ReadTodoRequest {
-    private Long userId;
-}
