--- src/main/java/com/ssafyss/board_practice/domain/todo/dto/UpdateTodoRequest.java
@@ -1,9 +1,0 @@
-package com.ssafyss.board_practice.domain.todo.dto;
-
-import lombok.Getter;
-
-@Getter
-public class UpdateTodoRequest {
-    private Long id;
-    private String content;
-}
