+++ src/main/java/com/ssafyss/board_practice/todo/dto/CreateTodoRequest.java
@@ -1,0 +1,9 @@
+package com.ssafyss.board_practice.todo.dto;
+
+import lombok.Getter;
+
+@Getter
+public class CreateTodoRequest {
+    private Long userId;
+    private String content;
+}
