+++ src/main/java/com/ssafyss/board_practice/todo/dto/DeleteTodoRequest.java
@@ -1,0 +1,8 @@
+package com.ssafyss.board_practice.todo.dto;
+
+import lombok.Getter;
+
+@Getter
+public class DeleteTodoRequest {
+    private Long id;
+}
