+++ src/main/java/com/ssafyss/board_practice/todo/dto/ReadTodoRequest.java
@@ -1,0 +1,8 @@
+package com.ssafyss.board_practice.todo.dto;
+
+import lombok.Getter;
+
+@Getter
+public class ReadTodoRequest {
+    private Long userId;
+}
