+++ src/main/java/com/ssafyss/board_practice/todo/dto/ReadTodoResponse.java
@@ -1,0 +1,14 @@
+package com.ssafyss.board_practice.todo.dto;
+
+import java.sql.Timestamp;
+import lombok.Getter;
+
+@Getter
+public class ReadTodoResponse {
+    private Long id;
+    private Long userId;
+    private String content;
+    private boolean completed;
+    private Timestamp createdAt;
+    private Timestamp updatedAt;
+}
