+++ src/main/java/com/ssafyss/board_practice/todo/dto/UpdateTodoRequest.java
@@ -1,0 +1,9 @@
+package com.ssafyss.board_practice.todo.dto;
+
+import lombok.Getter;
+
+@Getter
+public class UpdateTodoRequest {
+    private Long id;
+    private String content;
+}
