+++ src/main/java/com/ssafyss/board_practice/todo/entity/Todo.java
@@ -1,0 +1,16 @@
+package com.ssafyss.board_practice.todo.entity;
+
+import java.sql.Timestamp;
+import lombok.Builder;
+
+@Builder
+public class Todo {
+    private Long id;
+    private Long userId;
+    private String content;
+    private boolean completed;
+    private Timestamp createdAt;
+    private Timestamp deletedAt;
+    private Timestamp updatedAt;
+    private boolean deleted;
+}
