+++ src/main/resources/mapper/user.xml
@@ -1,14 +1,14 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.ssafyss.board_practice.domain.user.dao.UserDao">
+<mapper namespace="com.ssafyss.board_practice.user.infrastructure.UserRepository">
     <select id="countByEmail" resultType="int">
         select count(*)
         from User
         where email = #{email};
     </select>
 
-    <insert id="insert" parameterType="com.ssafyss.board_practice.domain.user.entity.User">
+    <insert id="insert" parameterType="com.ssafyss.board_practice.user.entity.User">
         insert into User (email, password)
         values (#{email}, #{password});
     </insert>
